port MinifyClosureCompilerTest to phpunit
