+++ min_unit_tests/test_all.php
@@ -9,7 +9,6 @@ require 'test_Minify_Cache_WinCache.php';
 require 'test_Minify_Cache_ZendPlatform.php';
 require 'test_Minify_CSS.php';
 require 'test_Minify_CSS_UriRewriter.php';
-require 'test_Minify_ClosureCompiler.php';
 require 'test_Minify_YuiCSS.php';
 require 'test_Minify_HTML.php';
 require 'test_Minify_ImportProcessor.php';
